+++ data/nextstrain/orthoebolavirus/sudv/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Align `pathogen.json` metadata with the current Nextclade schema layout.
+
 ## 2025-09-24T07:29:03Z
 
 Initial release of this dataset.